requirements.txt
numpy~=1.24.4
torch~=2.1.2+cu118
transformers~=3.4.0
py2neo~=2021.2.4
jieba~=0.42.1
scikit-learn~=1.3.2
tqdm~=4.66.1
pdfplumber~=0.11.0
pandas~=2.0.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
requests~=2.31.0
spacy~=3.7.2
flask~=3.0.2
protobuf~=3.20.0
grpcio~=1.62.1
grpcio-tools~=1.48.2